[F]: sqlRc

+++ internal/handlers/dto/transactions.go
@@ -3,11 +3,11 @@ package dto
 type TTransactionRs struct {
 	Description string `json:"description" validate:"required" `
 	Currency    string `json:"currency" validate:"required" `
-	ID          int64  `json:"id" validate:"required" `
-	OwnerID     int64  `json:"ownerID" validate:"required" `
-	Amount      int64  `json:"amount" validate:"required" `
-	IsIncome    int64  `json:"IsIncome" validate:"required" `
-	Createdat   int64  `json:"createdAt" validate:"required" `
+	ID          int    `json:"id" validate:"required" `
+	OwnerID     int    `json:"ownerID" validate:"required" `
+	Amount      int    `json:"amount" validate:"required" `
+	IsIncome    int    `json:"IsIncome" validate:"required" `
+	Createdat   int    `json:"createdAt" validate:"required" `
 }
 type TTransactionWTagsRs struct {
 	Tags []string `json:"tags" validate:"required"`
@@ -17,6 +17,6 @@ type TTransactionWTagsRs struct {
 type TTransactionRq struct {
 	Description string   `json:"description" validate:"required" `
 	Tags        []string `json:"tags" validate:"required"`
-	Amount      int64    `json:"amount" validate:"required" `
-	IsIncome    int64    `json:"isIncome" validate:"required" `
+	Amount      int      `json:"amount" validate:"required" `
+	IsIncome    int      `json:"isIncome" validate:"required" `
 }
